Added all lastest changes:
	- Episodes are now in a separate controller.
	- Seasons will be added in the Show's Controller
	- Working Proposals for new Show's
	- Added new Show Data on ShowMemory Repository.

Note: I dont like long commits, but i'm too tired to make it a 'foreach( Feature feature in NewFeatures)'

+++ MVC3/FollowMyTv.DomainLayer/Episode.cs
@@ -3,9 +3,13 @@ namespace FollowMyTv.DomainLayer
     public class Episode
     {
         public int Number { get; set; }
+
         public string Title { get; set; }
-        public long Duration { get; set; }
+
+        public int Duration { get; set; }
+
         public string Synopsis { get; set; }
+
         public int Score { get; set; }
     }
 }
